chore(circle-progress): split demo
