+++ packages/nutui-taro-demo/src/exhibition/pages/circleprogress/basic.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-circle-progress :progress="20" />
+</template>
