+++ packages/nutui-taro-demo/src/exhibition/pages/circleprogress/content.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-circle-progress :progress="90"> 90! </nut-circle-progress>
+</template>
