+++ packages/nutui-taro-demo/src/exhibition/pages/circleprogress/size.vue
@@ -1,0 +1,7 @@
+<template>
+  <nut-space align="baseline">
+    <nut-circle-progress :progress="80" :radius="80" />
+    <nut-circle-progress :progress="50" :radius="50" />
+    <nut-circle-progress :progress="30" :radius="30" />
+  </nut-space>
+</template>
