+++ packages/nutui-taro-demo/src/exhibition/pages/circleprogress/width.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-circle-progress :progress="50" :stroke-width="10" />
+</template>
