+++ src/packages/__VUE/circleprogress/demo/basic.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-circle-progress :progress="20" />
+</template>
