+++ src/packages/__VUE/circleprogress/demo/index.vue
@@ -20,7 +20,7 @@
   </Demo>
 </template>
 <script setup lang="ts">
-import { useTranslate } from '../../../utils';
+import { useTranslate } from '@/sites/utils';
 import Basic from './basic.vue';
 import Width from './width.vue';
 import Color from './color.vue';
